Revert "Fix reader index to use size_t."
This reverts commit cbe6261ea6db306219fb4ed52d63fa2c9a9ce13d.

+++ src/dpfp/fp.cc
@@ -438,7 +438,7 @@ int DPFPReader::getReaderIdx()
     return mReaderIdx;
 }
 
-bool DPFPReader::setReaderIdx(size_t reader)
+bool DPFPReader::setReaderIdx(int reader)
 {
     if (reader >= 0 && reader < mReaders.size()) {
         mReaderIdx = reader;
